2dcontext/2d.getImageData.negOffset.sourceRect.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /2dcontext/tools/gentest.py. -->
<title>Canvas test: 2d.getImageData.negOffset.sourceRect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.getImageData.negOffset.sourceRect</h1>
<p class="desc">Test the handling of negative sourceRect offset in getImageData().</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test the handling of negative sourceRect offset in getImageData().");
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'red';
ctx.fillRect(0, 0, 100, 100);
ctx.fillStyle = 'green';
ctx.fillRect(0, 0, 6, 6);

var imageData = ctx.getImageData(-10, 0, 100, 1);
var imgdata = imageData.data;

_assert(imgdata[0] == 0, "imgdata[\""+(0)+"\"] == 0");
_assert(imgdata[1] == 0, "imgdata[\""+(1)+"\"] == 0");
_assert(imgdata[2] == 0, "imgdata[\""+(2)+"\"] == 0");
_assert(imgdata[3] == 0, "imgdata[\""+(3)+"\"] == 0");

_assert(imgdata[60] == 0, "imgdata[\""+(60)+"\"] == 0");
_assert(imgdata[61] == 128, "imgdata[\""+(61)+"\"] == 128");
_assert(imgdata[62] == 0, "imgdata[\""+(62)+"\"] == 0");
_assert(imgdata[63] == 255, "imgdata[\""+(63)+"\"] == 255");

_assert(imgdata[64] == 255, "imgdata[\""+(64)+"\"] == 255");
_assert(imgdata[65] == 0, "imgdata[\""+(65)+"\"] == 0");
_assert(imgdata[66] == 0, "imgdata[\""+(66)+"\"] == 0");
_assert(imgdata[67] == 255, "imgdata[\""+(67)+"\"] == 255");


});
</script>
